03_Products/0302_CreateBasicTables
- Creación de los archivos **`companies.sql`**, **`docTypes.sql`**, **`branches.sql`** y **`roles.sql`**.
- Corrijo el archivo **`users.sql`**.
- Ejecuto el contenido de cada archivo _sql_ en `supabase` .

#03_Products/0302_CreateBasicTables

+++ src/db/sql/tables/users.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS users (
     email VARCHAR(100) UNIQUE NOT NULL,
     password_hash VARCHAR(255) NOT NULL,
     name VARCHAR(100),
-    id_type INT8 NOT NULL,
+    id_type INT NOT NULL,
     document VARCHAR(20) UNIQUE NOT NULL,
     phone VARCHAR(15),
     id_role INT NOT NULL,
